Add files via upload
Soil toxicity prediction and recommendation system using data mining in precision agriculture
Author: Mayuri Pawar, Geetha Chillarge
 In this paper, the proposed system can help farmers by making them aware about their soil conditions. Farmers can maximize crops yield by knowing proportion of nutrients present in the soil. Soil toxicity affects the soil nutrients which indirectly affects crops health. The proposed system predicts the level of toxicity present in the soil and makes farmer aware about it. Many farmers are depending on rainfall which is the one of the factor for poor growth and decreases crops yield. Thus the proposed system recommends the farmer about the crop, fertility of soil, level of toxicity and water supply. For this recommendation system, sensor's accuracy is very important as well as classification algorithm. For classification, decision tree J48 algorithm is used which is simple to implement and having more accuracy as compared with other classification algorithms. Issue of power supply can be overcome by using solar panel system.
 
 
Prediction of crop yield and fertilizer recommendation using machine learning algorithms
Author: Devdatta A Bondre, Santosh Mahagaonkar
The yield prediction is a major issue that remains to be solved based on available data. Machine learning techniques are the better choice for this purpose. Different Machine learning techniques are used and evaluated in agriculture for estimating the future year's crop production. This paper proposes and implements a system to predict crop yield from previous data. This is achieved by applying machine learning algorithms like Support Vector Machine and Random Forest on agriculture data and recommends fertilizer suitable for every particular crop. The paper focuses on creation of a prediction model which may be used for future prediction of crop yield. It presents a brief analysis of crop yield prediction using machine learning techniques.
 
Soil Based Fertilizer Recommendation System for Crop Disease Prediction System
Author :P Pandi Selvi, P Poornima
The proposed system was organized in such a way, to analyze the soil type, diseases in the leaves and finally to recommend the appropriate fertilizer to the farmers, that may be of great help to them.Smart analysis and Comprehensive prediction model in agriculture helps the farmer to yield right crop at the right time. The main benefits of the proposed system are as follows: Yield right crop at the right time, Balancing the crop production, control plant disease, Economic growth, and planning to reduce the crop scarcity. Hence to Detect and recognize the plant diseases and to recommend fertilizer it is necessary to provide symptoms in identifying the disease at its earliest. Hence the authors proposed and implemented new fertilizers Recommendation System for crop disease prediction.
 

KRISHI RAKSHAN-A Machine Learning based New Recommendation System to the Farmer
Author: DNVSLS Indira, M Sobhana, AHL Swaroop, V Phani Kumar
The main purpose of this research work is to develop a machine learning-based recommendation system to increase agricultural productivity. A variety of datasets were used in this study to design and develop advanced models to estimate the crop, recommend fertiliser, and identify plant disease. An algorithm called MobileNet uses an image of a leaf to identify the disease present in a plant. The XGBoost model predicts a suitable crop based on the local soil nutrients and rainfall. Random Forest [RF] model was used to propose fertilizer and develop ideas for improving soil fertility depending on nutrients present in the soil. When compared to other approaches, the proposed model delivers a high level of accuracy. Moreover, this article suggests the farmer to increase the crop yield by entering the input values and local soil conditions, wherein the model suggests recommended crop for that soil with an accuracy of 99%.
 
Recommendation System for Agriculture Using Machine Learning and Deep Learning
Author: K SuriyaKrishnaan, L Charan Kumar, R Vignesh
 
Picking the right crop for the land, cultivating it and obtaining a prosperous yield with the right fertilizer is a great challenge. The proposed system recommends the suitable crops for the lands with varied soil nutrients. The appropriate fertilizers that are suitable for specific soil nutrient and crop sown are also recommended. Plant physiology can be damaged due to fungal, viral or bacterial diseases. Plants affected from the above pathogens are detected. Random forest classifier gives an accuracy of 98% for recommendation system, and PyTorch neural network gives an accuracy of 99.2% for disease prediction.
 
Agriculture Based Recommendation System with Image Processing
Author: K Saranya, Deena Dhayalan, R Prasanth, M Sathish
Agriculture is the backbone of India as it plays a major role in Employment and Economy. One of the main reasons for loss in Agriculture is poor selection of crops that are to be grown. Most of the farmers are also not aware of requirements of soil like Nutrients, Minerals, Moisture content and others. This causes mental and financial stress to farmers. Other Major problem that a farmer faces is the disease and pest that affects the plant, which are aware only in later stages. To get better of this scenario, a model is suggested which recommends the most suitable crop by considering parameters like weather and soil based on live location. Along with this another model is constructed to predict the disease and suggest pesticides for that disease.
 



Crop Recommendation System using Machine Learning
Author: Dhruvi Gosai, Chintal Raval, Rikin Nayak, Hardik Jayswal, Axat Patel
 
A vast fraction of the population of India considers agriculture as its primary occupation. The production of crops plays an important role in our country. Bad quality crop production is often due to either excessive use of fertilizer or using not enough fertilizer. The proposed system of IoT and ML is enabled for soil testing using the sensors, is based on measuring and observing soil parameters. This system lowers the probability of soil degradation and helps maintain crop health. Different sensors such as soil temperature, soil moisture, pH, NPK, are used in this system for monitoring temperature, humidity, soil moisture, and soil pH along with NPK nutrients of the soil respectively. The data sensed by these sensors is stored on the microcontroller and analyzed using machine learning algorithms like random forest based on which suggestions for the growth of the suitable crop are made. This project also has a methodology that focuses on using a convolutional neural network as a primary way of identifying if the plant is at risk of a disease or not.